backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-extensions==3.2.3
whitenoise==6.6.0
dj-database-url==2.1.0
boto3==1.34.14
django-storages==1.14.2
